frame: Add a TODO item
@tdaede pointed out that the list is:
- Incomplete
- Not completely useful as it is

To most encoders the useful infomation is just if the frame to encode
should close the current group of pictures or not.
Few encoders might make use of the precise information regarding what
to do with the frame assuming the correct kind of frame matches.
Some encoders might take the hint to make the frame optionally
decodable.