Remove `WriteSeek` and `WriteOwned` traits
The fact that these traits require `ToOwned` made them too restrictive
to be used with real-world writers such as `File` and `Stdout`. However,
this did break some usages with `Cursor`s. The way those cursors were
being used was an anti-pattern, however, and cursors are still usable
with this package if taken by reference instead of moved.